# doc/api.yml
openapi: '3.0.2'
info:
  title: PiBank API
  version: 'latest'
servers:
  - url: https://api.pibank.test
tags:
  - name: User
  - name: Account
  - name: Balance
  - name: Transfer
paths:
  /user:
    get:
      tags:
        - "User"
      summary: "get all user"
      operationId: getAllUser
      responses:
        '200':
          description: successfully retrieved Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        '400':
          $ref: "#/components/responses/400"
        '500':
          $ref: "#/components/responses/500"
    put:
      tags:
        - "User"
      summary: create or update a user
      operationId: CrupdateUser
      requestBody:
        description: user to crupdate
        required: true
        content:
          application/json:
            schema:
              type: object
              items:
                $ref: "#/components/schemas/User"
      responses:
        '200':
          description: successfully crupdate Users
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
        '400':
          $ref: "#/components/responses/400"
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: "#/components/responses/500"
  /user/{id}:
    get:
      tags:
        - "User"
      summary: "get user by  id"
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - "User"
      summary: update a user by  id
      operationId: updateUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: user to update
        required: true
        content:
          application/json:
            schema:
              type: object
              items:
                $ref: "#/components/schemas/User"
      responses:
        '200':
          description: successfully crupdate Users
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
        '400':
          $ref: "#/components/responses/400"
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: "#/components/responses/500"
    delete:
      tags:
        - "User"
      summary: "delete user by  id"
      operationId: deleteUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /account:
    get:
      tags:
        - "Account"
      summary: "get all account"
      operationId: getAllAccount
      responses:
        '200':
          description: successfully retrieved Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        '400':
          $ref: "#/components/responses/400"
        '500':
          $ref: "#/components/responses/500"
    put:
      tags:
        - "Account"
      summary: create or update a account
      operationId: CrupdateUser
      requestBody:
        description: account to crupdate
        required: true
        content:
          application/json:
            schema:
              type: object
              items:
                $ref: "#/components/schemas/Account"
      responses:
        '200':
          description: successfully crupdate Accounts
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Account"
        '400':
          $ref: "#/components/responses/400"
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: "#/components/responses/500"
  /account/{accountNumber}:
    get:
      tags:
        - "Account"
      summary: "get account by  account number"
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - "Account"
      summary: update account by account number
      operationId: updateAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: account to update
        required: true
        content:
          application/json:
            schema:
              type: object
              items:
                $ref: "#/components/schemas/Account"
      responses:
        '200':
          description: successfully update account
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Account"
        '400':
          $ref: "#/components/responses/400"
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: "#/components/responses/500"
    delete:
      tags:
        - "Account"
      summary: "delete user by account number"
      operationId: deleteAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /account/{userId}/balance:
    get:
      tags:
        - "Account"
      summary: "get balance by userID"
      operationId: getBalanceByUserId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'

  /transfer:
    get:
      tags:
        - "Transfer"
      summary: "get all transfer"
      operationId: getAllTransfer
      responses:
        '200':
          description: successfully retrieved Transfer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"
        '400':
          $ref: "#/components/responses/400"
        '500':
          $ref: "#/components/responses/500"
    post:
      tags:
        - "Transfer"
      summary: create transfer
      operationId: CreateTransfer
      requestBody:
        description: balance to transfer
        required: true
        content:
          application/json:
            schema:
              type: object
              items:
                $ref: "#/components/schemas/Transfer"
      responses:
        '200':
          description: successfully create Transfer
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Transfer"
        '400':
          $ref: "#/components/responses/400"
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: "#/components/responses/500"
  /transfer/{userId}:
    get:
      tags:
        - "Transfer"
      summary: "get transfer by  id"
      operationId: getTransferById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - "Transfer"
      summary: update transfer by  id
      operationId: updateTransferById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: transfer to update
        required: true
        content:
          application/json:
            schema:
              type: object
              items:
                $ref: "#/components/schemas/Transfer"
      responses:
        '200':
          description: successfully update transfer
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Transfer"
        '400':
          $ref: "#/components/responses/400"
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: "#/components/responses/500"
    delete:
      tags:
        - "Transfer"
      summary: "delete transfer by id"
      operationId: deleteTransferById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'

components:
  schemas:
    Bank:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "BMOI"
        reference:
          type: string
          example: "004"
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        birthdayDate:
          type: string
          format: date
        netMonthSalary:
          type: number
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountNumber:
          type: number
        creditAllow:
          type: number
        overDraftLimit:
          type: boolean
        mainBalance:
          type: number
        loans:
          type: number
        interestLoans:
          type: number
        user:
          allOf:
            - $ref: "#/components/schemas/User"
        bank:
          allOf:
            - $ref: "#/components/schemas/Bank"
    BalanceHistory:
      type: object
      properties:
        mainBalance:
          type: number
        loans:
          type: number
        loansInterest:
          type: number
        date:
          type: string
          format: date
        Account:
          allOf:
            - $ref: "#/components/schemas/Account"
    Transaction:
      type: object
      properties:
        label:
          type: string
        type:
            type: string
            enum:
              - DEBIT
              - CREDIT
    Transfer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        transferReason:
          type: string
        amount:
          type: number
        effectiveDate:
          type: string
          format: date
        registrationDate:
          type: string
          format: date
        isCanceled:
          type: boolean
          default: false

    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'