# src/main/java/com/devyforu/pibanks/Model/Transaction.java
package com.devyforu.pibanks.Model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class Transaction {
    private String id;
    private Transfer Transfer;
    private String label;
    private TransactionType type;
}